+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fixed scatter chart for Highcharts
 - Limit concurrent requests to avoid overloading servers
 - Added ability to specify color with series for Chart.js and Highcharts
+- Added `round` option
 - Added bubble chart for Chart.js [experimental]
 
 ## 2.2.1